Link against and include the CUDA runtime library for C++ code, main branch (2021.03.01.)

+++ cuda/src/memory/cuda/arena_memory_manager.cpp
@@ -7,7 +7,7 @@
 
 // Local include(s).
 #include "vecmem/memory/cuda/arena_memory_manager.hpp"
-#include "vecmem/utils/cuda_error_handling.hpp"
+#include "../../utils/cuda_error_handling.hpp"
 
 // CUDA include(s).
 #include <cuda_runtime.h>
@@ -115,7 +115,8 @@ namespace vecmem { namespace cuda {
       // Increment the internal pointer.
       mem.m_nextAllocation += sizeInBytes + padding;
       // And make sure that we didn't run out of memory.
-      if( mem.m_nextAllocation - mem.m_ptr >= mem.m_size ) {
+      if( static_cast< std::size_t >( mem.m_nextAllocation - mem.m_ptr ) >=
+          mem.m_size ) {
          throw std::bad_alloc();
       }
